refactor: removed temperature method to work with api data

+++ lib/core/domain/weather/entities/core_weather_info_entity.dart
@@ -10,7 +10,7 @@ class CoreWeatherInfo {
     this.pressureGroundLevel,
   });
 
-  final double? temp;
+  final int? temp;
   final double? tempSensation;
   final double? tempMin;
   final double? tempMax;
